pyproject.toml
[project]
name = "2-wheelchair-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pytz>=2025.2",
    "tabulate>=0.9.0",
    "Flask>=3.0.3",
    "Flask-Cors>=4.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]
